Handle error on higher tree level
