+++ src/app.tsx
@@ -24,8 +24,6 @@ import ScssVars from "./assets/styles/styles/custom/_variables.module.scss";
 import { ApplicationStateModel } from "./state/models/application-state.model";
 import { ScheduleKey } from "./api/persistance-store.model";
 import { AppMode, useAppConfig } from "./state/app-config-context";
-import * as Sentry from "@sentry/react";
-import AppErrorModal from "./components/common-components/modal/app-error-modal/app-error.modal.component";
 import { cropScheduleDMToMonthDM } from "./logic/schedule-container-convertion/schedule-container-convertion";
 
 const useStyles = makeStyles(() => ({
@@ -104,39 +102,25 @@ function App(): JSX.Element {
     fetchGlobalState();
   }, [fetchGlobalState]);
 
-  const [open, setIsOpen] = useState(false);
-  const fallback = useCallback(
-    ({ resetError }): JSX.Element => (
-      <AppErrorModal onClick={resetError} open={open} setOpen={setIsOpen} />
-    ),
-    [open, setIsOpen]
-  );
-
-  const onError = useCallback((): void => {
-    setIsOpen(true);
-  }, [setIsOpen]);
-
   return (
-    <Sentry.ErrorBoundary fallback={fallback} onError={onError}>
-      <NotificationProvider>
-        <JiraLikeDrawerProvider>
-          <Switch>
-            <Route path="/">
-              <Box className={classes.root}>
-                <Box className={classes.content}>
-                  <HeaderComponent />
-                  <RouteButtonsComponent tabs={tabs} disabled={disableRouteButtons} />
-                </Box>
-                <Box className={classes.drawer}>
-                  <JiraLikeDrawer width={690} />
-                </Box>
+    <NotificationProvider>
+      <JiraLikeDrawerProvider>
+        <Switch>
+          <Route path="/">
+            <Box className={classes.root}>
+              <Box className={classes.content}>
+                <HeaderComponent />
+                <RouteButtonsComponent tabs={tabs} disabled={disableRouteButtons} />
               </Box>
-              {isElectron() ? <></> : <NetlifyProFooter />}
-            </Route>
-          </Switch>
-        </JiraLikeDrawerProvider>
-      </NotificationProvider>
-    </Sentry.ErrorBoundary>
+              <Box className={classes.drawer}>
+                <JiraLikeDrawer width={690} />
+              </Box>
+            </Box>
+            {isElectron() ? <></> : <NetlifyProFooter />}
+          </Route>
+        </Switch>
+      </JiraLikeDrawerProvider>
+    </NotificationProvider>
   );
 }
